pyproject.toml
[project]
name = "keycloak-mcp-server"
version = "1.0.0"
description = "A comprehensive MCP server for Keycloak Admin REST API"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
sse = [
    "uvicorn>=0.30.0",
    "starlette>=0.38.0",
]

[project.scripts]
keycloak-mcp-server = "keycloak_mcp_server.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/keycloak_mcp_server"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0",
    "pyright>=1.1.300",
]
